Criação de component e correção de css
